requirements.txt
python-dotenv
google-generativeai
google-search-results
supabase